Propper browser persistent cache (#36)
This re-implements the browser caching behavior by using IDB and not saving the key/values directly but instead (non-blocking) saving the entire hashmap and using it for (non-blocking) rehydration on page load. This solves all the performance issues of the previous implementation.

The catch is that the rehydration sometimes takes up to 2 seconds, so if API requests are made they will go out during that. This is generally OK because it ends up being the fastest experience for the user 👍